Fix: Round face clip
Fix: Resolve conflicts

Fix: Refactor functino call

Fix: Consider tiny faces

Fix: Add comment for function

Fix: Refactor vertex structure
