+++ src/app/three-extensions/ThreeUtils.js
@@ -282,7 +282,7 @@ const ThreeUtils = {
 
         function isSimilarPlanes(p1, p2) {
             return Math.abs(p1.distanceToPoint(center) - p2.distanceToPoint(center)) < 0.05
-            && p1.normal.angleTo(p2.normal) * 180 / Math.PI < 1;
+            && p1.normal.angleTo(p2.normal) * 180 / Math.PI < 0.6;
         }
 
         const positions = geometry.getAttribute('position').array;
@@ -301,7 +301,7 @@ const ThreeUtils = {
             triangle.getPlane(plane);
             const area = triangle.getArea();
             // skip tiny triangles
-            if (area < 0.1) continue;
+            // if (area < 0.1) continue;
 
             if (baseMode) {
                 if (inverseNormal) {
@@ -333,6 +333,115 @@ const ThreeUtils = {
         }
 
         return { planes, areas, supportVolumes, planesPosition };
+    },
+
+    isReverseEdge(e1, e2) {
+        return e1[0].x === e2[1].x && e1[0].y === e2[1].y && e1[0].z === e2[1].z
+        && e2[0].x === e1[1].x && e2[0].y === e1[1].y && e2[0].z === e1[1].z;
+    },
+
+    checkEdgeAvailable(newEdge, edges) {
+        let foundReverseEdge = false;
+        for (let j = edges.length - 1; j > -1; j--) {
+            const edge = edges[j];
+            if (this.isReverseEdge(edge, newEdge)) {
+                foundReverseEdge = true;
+                edges.splice(j, 1);
+            }
+        }
+        if (!foundReverseEdge) {
+            edges.push(newEdge);
+        }
+    },
+
+    arrangeEdges(edges) {
+        let edge = edges[0];
+        const tempEdges = edges.slice(1);
+        const arrangedEdges = [edge];
+        while (tempEdges.length > 0) {
+            for (let j = tempEdges.length - 1; j > -1; j--) {
+                const edgeChecked = tempEdges[j];
+                if (edge[1].x === edgeChecked[0].x && edge[1].y === edgeChecked[0].y && edge[1].z === edgeChecked[0].z) {
+                    arrangedEdges.push(edgeChecked);
+                    edge = edgeChecked;
+                    tempEdges.splice(j, 1);
+                    break;
+                }
+            }
+        }
+        return arrangedEdges;
+    },
+
+    generateRotationFaces(positions, modelBoundingBox) {
+        // extract points
+        const vertex = [];
+        for (let i = 0; i < positions.length; i += 3) {
+            vertex.push(new THREE.Vector3(
+                positions[i],
+                positions[i + 1],
+                positions[i + 2]
+            ));
+        }
+
+        // delete reverse edges
+        let edges = [];
+        for (let i = 0; i < vertex.length; i += 3) {
+            const edgeA = [vertex[i], vertex[i + 1]];
+            const edgeB = [vertex[i + 1], vertex[i + 2]];
+            const edgeC = [vertex[i + 2], vertex[i]];
+            this.checkEdgeAvailable(edgeA, edges);
+            this.checkEdgeAvailable(edgeB, edges);
+            this.checkEdgeAvailable(edgeC, edges);
+        }
+        // connect edges to a polyline
+        edges = this.arrangeEdges(edges);
+
+        const facePos = [];
+        /**
+         * shrink the polyline area
+         *       / \
+         *     / /\ \
+         *   / /___\ \
+         * /__________\
+         */
+        function genFacePos(p1, p2, p3) {
+            const n = 2;
+            const x = (p1.x + 6 * p2.x + p3.x) / (4 * n);
+            const y = (p1.y + 6 * p2.y + p3.y) / (4 * n);
+            const z = (p1.z + 6 * p2.z + p3.z) / (4 * n);
+            facePos.push(new THREE.Vector3(x, y, z));
+        }
+        for (let i = 0; i < edges.length - 1; i++) {
+            const point1 = edges[i][0];
+            const point2 = edges[i][1];
+            const point3 = edges[i + 1][1];
+            genFacePos(point1, point2, point3);
+        }
+        genFacePos(edges[edges.length - 1][0], edges[edges.length - 1][1], edges[0][1]);
+
+        // generate curve line
+        const curve = new THREE.CatmullRomCurve3(facePos, true, 'catmullrom');
+        const points = curve.getPoints(50);
+        const curveFaceVertex = [points[0], points[1], points[2]];
+        for (let i = 3; i < points.length; i++) {
+            const last = curveFaceVertex[curveFaceVertex.length - 1];
+            curveFaceVertex.push(curveFaceVertex[0]);
+            curveFaceVertex.push(last);
+            curveFaceVertex.push(points[i]);
+        }
+
+        // clip vertex outside model boundingbox
+        for (let i = 0; i < curveFaceVertex.length; i++) {
+            curveFaceVertex[i].x = (curveFaceVertex[i].x > modelBoundingBox.max.x ? modelBoundingBox.max.x : curveFaceVertex[i].x);
+            curveFaceVertex[i].x = (curveFaceVertex[i].x < modelBoundingBox.min.x ? modelBoundingBox.min.x : curveFaceVertex[i].x);
+
+            curveFaceVertex[i].y = (curveFaceVertex[i].y > modelBoundingBox.max.y ? modelBoundingBox.max.y : curveFaceVertex[i].y);
+            curveFaceVertex[i].y = (curveFaceVertex[i].y < modelBoundingBox.min.y ? modelBoundingBox.min.y : curveFaceVertex[i].y);
+
+            curveFaceVertex[i].z = (curveFaceVertex[i].z > modelBoundingBox.max.z ? modelBoundingBox.max.z : curveFaceVertex[i].z);
+            curveFaceVertex[i].z = (curveFaceVertex[i].z < modelBoundingBox.min.z ? modelBoundingBox.min.z : curveFaceVertex[i].z);
+        }
+        return curveFaceVertex;
     }
 };
